Made a toolbar
