+++ src/components/WebApp/AddBookmarkSection/AddBookmarkBtnElements.js
@@ -2,12 +2,16 @@ import styled from 'styled-components';
 
 export const Btn = styled.h1`
     padding: 20px;
-    background: gray;
-    border-radius: 6px;
+    background: #ccc5c5e7;
+    border-radius: 7px;
     cursor: pointer;
     text-align: center;
     text-transform: uppercase;
     width: 300px;
     margin: 20px auto 30px;
     user-select: none;
+
+    &:hover {
+        background: #ffffffe6;
+    }
 `;
